# Conflicts:
#	src/main/java/org/example/pmas/service/ProjectService.java

+++ src/main/java/org/example/pmas/service/ProjectService.java
@@ -1,12 +1,15 @@
 package org.example.pmas.service;
 
-import org.example.pmas.model.dto.ProjectDTO;
+import org.example.pmas.dto.ProjectDTO;
+import org.example.pmas.dto.SubProjectDTO;
 import org.example.pmas.exception.NotFoundException;
 import org.example.pmas.model.Project;
 import org.example.pmas.model.User;
 import org.example.pmas.repository.Interfaces.IProjectRepository;
 import org.example.pmas.repository.Interfaces.ISubProjectRepository;
 import org.example.pmas.repository.Interfaces.IUserRepository;
+import org.example.pmas.repository.UserRepository;
+import org.example.pmas.service.comparators.ProjectDTODeadlineComparator;
 import org.example.pmas.service.comparators.ProjectDeadlineComparator;
 import org.example.pmas.util.CompletionStatCalculator;
 import org.springframework.stereotype.Service;
@@ -14,18 +17,17 @@ import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Set;
 
 @Service
 public class ProjectService {
 
     private final IProjectRepository projectRepository;
-    private final ISubProjectRepository subprojectRepository;
     private final IUserRepository userRepository;
 
 
-    public ProjectService(IProjectRepository projectRepository, ISubProjectRepository subprojectRepository, IUserRepository userRepository) {
+    public ProjectService(IProjectRepository projectRepository, IUserRepository userRepository) {
         this.projectRepository = projectRepository;
-        this.subprojectRepository = subprojectRepository;
         this.userRepository = userRepository;
 
     }
@@ -34,12 +36,18 @@ public class ProjectService {
         return projectRepository.create(project);
     }
 
-    public void addUsersToProject(int projectID, List<Integer> userIDs){
-        projectRepository.addUsersToProject(projectID, userIDs);
+    public void addUsersToProject(int projectID, Set<Integer> userIDs){
+        if(userIDs != null){
+            projectRepository.addUsersToProject(projectID, userIDs);
+        }
+
     }
 
-    public void removeUsersFromProject(int projectID, List<Integer> userIds){
-        projectRepository.removeUsersFromProject(projectID, userIds);
+    public void removeUsersFromProject(int projectID, Set<Integer> userIds){
+        if(userIds != null){
+            projectRepository.removeUsersFromProject(projectID, userIds);
+        }
+
     }
 
     public List<Project> readAll(){
@@ -90,27 +98,34 @@ public class ProjectService {
     }
 
     public List<ProjectDTO> getProjectDTOByUserID(int userID){
+        // Returns null if no list
         List<ProjectDTO> projects = projectRepository.getProjectDTOByUserID(userID);
+        if(projects == null) return Collections.emptyList();
 
         for(ProjectDTO p : projects) {
             p.setCompletionPercentage(CompletionStatCalculator.calculatePercentage(p.getCompletedSubProjects(),p.getTotalSubProjects()));
             p.setCompleted(CompletionStatCalculator.isJobCompleted(p.getCompletedSubProjects(), p.getTotalSubProjects()));
         }
 
-        return projects;
+        // Sort the list by deadline
+        // We copy the list so its not immutable
+        List<ProjectDTO> modifiableList = new ArrayList<>(projects);
+        modifiableList.sort(new ProjectDTODeadlineComparator().reversed());
+        return modifiableList;
     }
 
     public List<User> getAllUsers(){
         return userRepository.readAll();
     }
 
-    // Sorts the list by deadline and then priority.
+    // Sorts the list by deadline.
     // If the list is null, return an empty list. No errors
     private List<Project> sortList(List<Project> projects){
         // If the list is null, return an empty list. No errors
         if(projects == null) return Collections.emptyList();
 
-        // Sort the list by deadline and then priority.
+        // Sort the list by deadline.
+        // We copy the list so its not immutable
         List<Project> modifiableList = new ArrayList<>(projects);
         modifiableList.sort(new ProjectDeadlineComparator().reversed());
